+++ Plugins/Rider/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.2] - 2024-01-21
+### Changed
+- Support 2024.1 EAP.
+- Wrap long lines in output.
+- Detect missing/outdated dotnet tool and offer install/upgrade. 
+
 ## [1.0.1] - 2023-12-16
 ### Changed
 - Minor fixes.